Added @testing-library/jest-dom and adjusted config files to support unit tests for front-end components. Wrote a simple test in App.test.jsx to make sure everything is working. Will rewrite next as I get into TDD for first player-facing feature.
