+++ package.json
@@ -33,10 +33,12 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/react": "^14.1.2",
     "@babel/preset-react": "^7.27.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^14.1.2",
     "@vitejs/plugin-react": "^4.4.1",
     "cypress": "^13.6.4",
+    "vite": "^6.3.5",
     "vitest": "^3.1.3"
   }
 }
