+++ vite.config.js
@@ -2,5 +2,9 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
 export default defineConfig({
+  test: {
+    globals: true,
+    setupFiles: 'src/setupTests.js'
+  },
   plugins: [react()],
 });
