+++ vitest.config.js
@@ -7,6 +7,7 @@ dotenv.config({ path: '.env.test' });
 export default defineConfig({
   plugins: [react()],
   test: {
+    setupFiles: 'src/setupTests.js',
     environment: 'jsdom',
     globals: true,
     exclude: ['.netlify/**', 'node_modules/**', 'dist/**']
